@@ -9,32 +9,32 @@
     "dev": "vite"
   },
   "dependencies": {
-    "@mdi/font": "7.3.67",
+    "@mdi/font": "7.4.47",
     "@vuelidate/core": "^2.0.3",
     "@vuelidate/validators": "^2.0.4",
-    "axios": "^1.6.1",
-    "core-js": "^3.33.2",
+    "axios": "^1.6.8",
+    "core-js": "^3.37.1",
     "roboto-fontface": "*",
-    "vue": "^3.3.8",
-    "vue-router": "^4.2.5",
-    "vuetify": "^3.4.0",
+    "vue": "^3.4.27",
+    "vue-router": "^4.3.2",
+    "vuetify": "^3.6.6",
     "vuex": "^4.1.0",
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.3",
-    "@babel/eslint-parser": "^7.23.3",
-    "@fortawesome/fontawesome-free": "^6.4.2",
-    "@vitejs/plugin-vue": "^4.4.1",
+    "@babel/core": "^7.24.5",
+    "@babel/eslint-parser": "^7.24.5",
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "@vitejs/plugin-vue": "^5.0.4",
     "@vue/cli-plugin-babel": "~5.0.8",
     "@vue/cli-plugin-eslint": "~5.0.8",
     "@vue/cli-plugin-router": "~5.0.8",
     "@vue/cli-plugin-vuex": "~5.0.8",
     "@vue/cli-service": "~5.0.8",
-    "eslint": "^8.53.0",
-    "eslint-plugin-vue": "^9.18.1",
-    "vite": "^4.5.0",
-    "vite-plugin-vuetify": "^1.0.2",
+    "eslint": "^9.2.0",
+    "eslint-plugin-vue": "^9.26.0",
+    "vite": "^5.2.11",
+    "vite-plugin-vuetify": "^2.0.3",
     "vue-cli-plugin-vuetify": "~2.5.8"
   },
   "eslintConfig": {
