@@ -1,8 +1,7 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import vuetify from 'vite-plugin-vuetify'
-
-const path = require('path')
+import path from 'path'
 
 // https://vitejs.dev/config/
 export default defineConfig({
